Moved report header to common components in folder
